+++ src/save.c
@@ -537,7 +537,7 @@ int makeSaveFile(int entry)
   FILE* fHandle;
   char buffer[100];
   int i;
-  unsigned long int var28 = 0;
+  u32 var28 = 0;
   char* ptr;
   int var_14 = 0;
   int temp;
